Remove broken `--force` option from `heroku plugins:link`
CI is showing an error. See the change between these two builds:
* https://travis-ci.org/heroku/heroku-ps-exec/jobs/368155915
* https://travis-ci.org/heroku/heroku-ps-exec/jobs/380432720
The first does not show the error, the second does. The Heroku CLI version is visible higher up in the build log - the earlier build got `heroku-cli/6.16.13-dbb9c23`, the latter `heroku/7.0.47`, so I assume it was removed between those two versions.

The effect of the `plugins:link` command failing is that the CI build was not actually testing the version of the plugin in the branch, it was testing the version bundled with the Heroku CLI version that was installed in the test setup. So I guess that means any changes _were_ eventually tested, but only once they'd been released!

The `errexit` bash option will be added in a later commit to make errors like this cause the CI build to fail.