Fix travis build and make it more reliable (#21)
* Enable "unofficial bash strict mode" at top of test scripts

To ensure we catch more bugs (see http://redsymbol.net/articles/unofficial-bash-strict-mode/). A couple changes were required to make this work, although I don't think any of them were causing any real problems yet:

* Change how we pass "yes" to `heroku keys:add`. As it stands this line will cause a "pipefail" error - the `yes` command will exit with a 141 ("terminated by a SIGPIPE") when the `heroku keys:add` command exits. Turns out `keys:add` has a `--yes` option so we can just use that instead.
* Give bash variables suitable default values. Otherwise they cause `nounset` errors.

* Remove broken `--force` option from `heroku plugins:link`

CI is showing an error. See the change between these two builds:
* https://travis-ci.org/heroku/heroku-ps-exec/jobs/368155915
* https://travis-ci.org/heroku/heroku-ps-exec/jobs/380432720
The first does not show the error, the second does. The Heroku CLI version is visible higher up in the build log - the earlier build got `heroku-cli/6.16.13-dbb9c23`, the latter `heroku/7.0.47`, so I assume it was removed between those two versions.

The effect of the `plugins:link` command failing is that the CI build was not actually testing the version of the plugin in the branch, it was testing the version bundled with the Heroku CLI version that was installed in the test setup. So I guess that means any changes _were_ eventually tested, but only once they'd been released!

The `errexit` bash option will be added in a later commit to make errors like this cause the CI build to fail.